requirements.txt
langchain-community
langchain
langchain-groq
langchain-openai
langchain-google-genai
PyPDF2
flask
dotenv-python
faiss-cpu
tiktoken
gunicorn==21.2.0
langchain-core